Extract reading game results from CSV to method
This will result in one extra iteration through the results but in this
case I think that it’s worth it because now we can keep all the logic
for reading from the CSV separate from building the leaderboard. We
also can reuse this new method to help with the refactoring I’m going
to do next for the team page.